split DemoSettingsSection into a separate view
